Trim id and url in parseResolver
