+++ modules/core/src/main/scala/org/scalasteward/core/buildsystem/maven/MavenAlg.scala
@@ -16,14 +16,11 @@
 
 package org.scalasteward.core.buildsystem.maven
 
-import atto.Atto._
-import atto._
 import better.files.File
 import cats.Monad
 import cats.implicits._
 import org.scalasteward.core.application.Config
 import org.scalasteward.core.buildsystem.BuildSystemAlg
-import org.scalasteward.core.buildsystem.maven.MavenParser._
 import org.scalasteward.core.data._
 import org.scalasteward.core.io.{FileAlg, ProcessAlg, WorkspaceAlg}
 import org.scalasteward.core.scalafix.Migration
@@ -51,8 +48,8 @@ object MavenAlg {
         listResolversCommand = mvnCmd(command.listRepositories)
         repositoriesRaw <- exec(listResolversCommand, repoDir)
         dependenciesRaw <- exec(listDependenciesCommand, repoDir)
-        (_, dependencies) = parseAllDependencies(dependenciesRaw)
-        (_, resolvers) = parseResolvers(repositoriesRaw.mkString("\n"))
+        (_, dependencies) = MavenParser.parseAllDependencies(dependenciesRaw)
+        (_, resolvers) = MavenParser.parseResolvers(repositoriesRaw.mkString("\n"))
       } yield {
         val deps = dependencies.distinct
         List(Scope(deps, resolvers))
@@ -73,87 +70,4 @@ object MavenAlg {
     def ignoreOptsFiles[A](dir: File)(fa: F[A]): F[A] =
       fileAlg.removeTemporarily(dir / ".jvmopts")(fa)
   }
-}
-
-object MavenParser {
-  private val dot: Parser[Char] = char('.')
-  private val underscore = char('_')
-  private val colon: Parser[Char] = char(':')
-
-  private val group: Parser[GroupId] = (for {
-    x <- (many1(noneOf(".:")) ~ opt(dot)).many1
-    _ <- colon
-  } yield {
-    //val parts: NonEmptyList[(NonEmptyList[Char], Option[Char])] = x
-    val groupVal = x.toList.map { case (g, d) => (g.toList ++ d.toList).mkString }.mkString
-    GroupId(groupVal)
-  }).named("group")
-
-  private val version: Parser[String] =
-    many1(noneOf(":")).map(_.mkString_(""))
-
-  private val artifact: Parser[ArtifactId] = {
-    val artifactString: Parser[String] = many1(noneOf("_ :")).map(_.toList.mkString)
-
-    for {
-      init <- artifactString
-      restOpt <- opt(many1(underscore ~ artifactString))
-      _ <- colon
-    } yield {
-      restOpt.fold(ArtifactId(init, Option.empty[String])) { rest =>
-        val crossVersion: Option[String] =
-          Option.when(rest.last._2.toFloatOption.isDefined)(rest.last._2)
-
-        val suffix = crossVersion.fold(
-          rest.map { case (underscore, str) => s"${underscore}${str}" }.toList.mkString
-        )(_ => rest.init.map { case (underscore, str) => s"${underscore}${str}" }.mkString)
-
-        ArtifactId(init + suffix, maybeCrossName = crossVersion)
-      }
-    }
-  }
-
-  private val configurations: Parser[Option[String]] =
-    string("compile").as(None) | string("test").map(Some(_))
-
-  private val parserDependency = for {
-    _ <- opt(string("[INFO]") <~ many(whitespace))
-    g <- group
-    a <- artifact
-    _ <- string("jar") <~ colon
-    v <- version <~ colon
-    c <- configurations
-  } yield {
-    Dependency(groupId = g, artifactId = a, version = v, configurations = c)
-  }
-
-  def parseAllDependencies(input: List[String]): (List[(String, String)], List[Dependency]) =
-    input
-      .map { line =>
-        val either = parserDependency.parse(line).done.either
-        either.leftMap(err => line -> err)
-      }
-      .partitionMap(identity)
-
-  val resolverIdParser: Parser[String] = many1(noneOf(" ")).map(_.toList.mkString)
-  val urlParser: Parser[String] = many1(noneOf(" ")).map(_.toList.mkString)
-
-  private val parserResolver = for {
-    _ <- many(whitespace)
-    _ <- string("id:") ~ whitespace
-    id <- resolverIdParser
-    _ <- many(whitespace) <~ string("url:") <~ whitespace
-    url <- resolverIdParser
-    _ <- many(anyChar)
-  } yield {
-    Resolver.MavenRepository(id, url, None)
-  }
-
-  def parseResolvers(raw: String): (List[String], List[Resolver.MavenRepository]) =
-    raw
-      .split("""\[INFO\]""")
-      .toList
-      .map(line => parserResolver.parse(line).done.either)
-      .partitionMap(identity)
-
 }
